--- src/main/java/racingcar/racing/model/Movable.java
@@ -1,5 +1,0 @@
-package racingcar.racing.model;
-
-public interface Movable {
-    void forward(boolean value);
-}
