--- src/main/java/racingcar/racing/policy/RacingMovePolicy.java
@@ -1,5 +1,0 @@
-package racingcar.racing.policy;
-
-public interface RacingMovePolicy {
-    boolean execute(int value);
-}
