+++ src/main/java/racingcar/v1/domain/Movable.java
@@ -1,0 +1,5 @@
+package racingcar.v1.domain;
+
+public interface Movable {
+    void forward(boolean value);
+}
