+++ src/main/java/racingcar/v1/domain/ParticipantName.java
@@ -1,0 +1,4 @@
+package racingcar.v1.domain;
+
+public class ParticipantName {
+}
