+++ src/main/java/racingcar/v1/domain/Position.java
@@ -1,0 +1,4 @@
+package racingcar.v1.domain;
+
+public class Position {
+}
