+++ src/main/java/racingcar/v1/domain/policy/RacingMovePolicy.java
@@ -1,0 +1,5 @@
+package racingcar.v1.domain.policy;
+
+public interface RacingMovePolicy {
+    boolean execute(int value);
+}
