+++ src/main/java/racingcar/v1/dto/RacingGameParam.java
@@ -1,0 +1,25 @@
+package racingcar.v1.dto;
+
+public class RacingGameParam {
+    private final int carNumber;
+    private final int tryNumber;
+    private final String[] names;
+
+    public RacingGameParam(int carNumber, int tryNumber, String[] names) {
+        this.carNumber = carNumber;
+        this.tryNumber = tryNumber;
+        this.names = names;
+    }
+
+    public int getCarNumber() {
+        return carNumber;
+    }
+
+    public int getTryNumber() {
+        return tryNumber;
+    }
+
+    public String[] getNames() {
+        return names;
+    }
+}
