+++ src/main/java/racingcar/v1/utils/validator/Validator.java
@@ -1,0 +1,5 @@
+package racingcar.v1.utils.validator;
+
+public interface Validator <T1, T2>{
+    T1 valid(T2 input);
+}
