+++ src/main/java/racingcar/v2/domain/policy/MovePolicy.java
@@ -1,0 +1,5 @@
+package racingcar.v2.domain.policy;
+
+public interface MovePolicy {
+    boolean apply(int value);
+}
