--- src/main/java/racingcar/validator/Validator.java
@@ -1,5 +1,0 @@
-package racingcar.validator;
-
-public interface Validator <T1, T2>{
-    T1 valid(T2 input);
-}
